stone/paper/scissors v1
This is my first project. Yes, I know it's not that hard. Yes, I know that it has not yet been completed, but that is why it is called v1. This is only the first version of my project. In the future, I'm going to add a repetition of the game here and so that if the spelling is incorrect, the game does not close.